backend/requirements.txt
asgiref==3.8.1
coloredlogs==15.0.1
Django==5.1.6
flatbuffers==25.2.10
humanfriendly==10.0
joblib==1.4.2
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.20.1
packaging==24.2
pandas==2.2.2
pillow==10.2.0
protobuf==5.29.3
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
sympy==1.13.3
threadpoolctl==3.5.0
tzdata==2025.1
requests==2.26.0
